Completado ejercicio 3 tema 9

+++ ejercicios_java/tema09/Tema09Ejercicio03AlbertoM/Canario.java
@@ -31,8 +31,8 @@ public class Canario extends Ave {
   }
 
   @Override
-  public int longitudAla(int envergadura) {
-    return envergadura / 2 - 3;
+  public double longitudAla(double envergadura) {
+    return (int) (envergadura / 2 - 3);
   }
 
   @Override
